+++ unique_toolkit/CHANGELOG.md
@@ -5,8 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), 
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.8.26] - 2025-08-28
+## [0.8.30] - 2025-08-28
 - Added A2A manager
+
+## [0.8.29] - 2025-08-27
+- Include `MessageExecution` and `MessageLog` in toolkit 
+
+## [0.8.28] - 2025-08-28
+- Fix paths for `sdk_url` and `openai_proxy` for localhost
+
+## [0.8.27] - 2025-08-28
+- Fixed function "create_async" in language_model.functions : "user_id" argument should be optional.
+
+## [0.8.26] - 2025-08-27
+- Optimized MCP manager
 
 ## [0.8.26] - 2025-08-27
 - Optimized MCP manager